perf, SEO, and a11y improvements + fix broken develop build
Follow-up to the i18n work. Groups several site improvements.

Build fix (develop is currently broken):
- Remove the stray pre-i18n src/app/become-partner/page.tsx that the PR #48
  merge reintroduced at the old route path (imports a missing CSS module and
  would create a non-localized /become-partner route). Live page stays at
  src/app/[lang]/become-partner.

Performance:
- Downscale/recompress oversized images in public/ (longest side ≤ 2000px,
  JPEG q72); image weight ~144MB → ~38MB. Remove unused public/dummy.jpg.
  (Annual-report PDFs in public/documents, ~211MB, left as a follow-up.)

SEO:
- app/sitemap.ts (both locales + all blog articles, with hreflang alternates)
  and app/robots.ts, using canonical origin https://www.oaninternational.org.
- metadataBase + Open Graph + Twitter card + og-image.jpg in the root layout;
  per-locale canonical/alternates/openGraph in [lang]/layout.
- Per-page localized generateMetadata for every route. Client pages are split
  into a server page.tsx (metadata) + page.client.tsx (body).
- Root "/" now emits a 0s meta-refresh + canonical (crawlers) alongside the
  JS redirect.
- postbuild script rewrites <html lang> to "en" in statically exported English
  pages, so crawlers get the correct lang without relying on client JS.

Accessibility:
- aria-labels on the mobile menu toggle, language switcher, and footer social
  links; aria-expanded / aria-current where relevant.
- Visible :focus-visible outline for interactive elements.

Consistency:
- Single MEMBERSHIP constant (current 84, target 100) used by the home hero,
  become-partner, and MembershipGoals. Standardizes the previously conflicting
  80/71 values to 84 per the team's latest "changed partners counter" change.
