+++ package.json
@@ -5,6 +5,7 @@
   "scripts": {
     "dev": "next dev",
     "build": "next build",
+    "postbuild": "node scripts/fix-html-lang.mjs",
     "start": "next start",
     "lint": "next lint"
   },
