+++ src/app/[lang]/layout.tsx
@@ -6,10 +6,16 @@ import Footer from "@/components/footer/footer.component";
 import HtmlLangSetter from "@/components/html-lang-setter/html-lang-setter.component";
 import { isLocale, locales, Locale } from "@/i18n/config";
 import { getDictionary } from "@/i18n/dictionaries";
+import { SITE_URL } from "@/i18n/routes";
 
 export function generateStaticParams() {
   return locales.map((lang) => ({ lang }));
 }
+
+const OG_LOCALE: Record<Locale, string> = {
+  es: "es_ES",
+  en: "en_US",
+};
 
 export async function generateMetadata({
   params,
@@ -22,9 +28,26 @@ export async function generateMetadata({
   }
   const dict = getDictionary(lang);
   return {
-    title: dict.metadata.title,
-    keywords: ["Nikki", "Benin", "ONG", "Nikarit"],
+    // Override the root template for this locale subtree; pages set only their
+    // own title and get "%s | OAN International".
+    title: {
+      default: dict.metadata.title,
+      template: `%s | ${dict.metadata.title}`,
+    },
+    keywords: ["Nikki", "Benin", "ONG", "NGO", "Nikarit", "OAN International"],
     description: dict.metadata.description,
+    alternates: {
+      canonical: `${SITE_URL}/${lang}`,
+      languages: Object.fromEntries(
+        locales.map((l) => [l, `${SITE_URL}/${l}`])
+      ),
+    },
+    openGraph: {
+      locale: OG_LOCALE[lang],
+      title: dict.metadata.title,
+      description: dict.metadata.description,
+      url: `${SITE_URL}/${lang}`,
+    },
   };
 }
