+++ src/app/[lang]/support-us/campaigns/grano-de-arena/page.tsx
@@ -1,6 +1,30 @@
+import type { Metadata } from "next";
 import { notFound } from "next/navigation";
 import styles from "./page.module.css";
 import { isLocale, Locale } from "@/i18n/config";
+
+const meta: Record<Locale, { title: string; description: string }> = {
+  es: {
+    title: "Grano de Arena",
+    description:
+      "Campaña Grano de Arena de OAN International para apoyar el desarrollo en Nikki.",
+  },
+  en: {
+    title: "Grano de Arena",
+    description:
+      "OAN International's Grano de Arena campaign supporting development in Nikki.",
+  },
+};
+
+export async function generateMetadata({
+  params,
+}: {
+  params: Promise<{ lang: string }>;
+}): Promise<Metadata> {
+  const { lang } = await params;
+  if (!isLocale(lang)) return {};
+  return meta[lang];
+}
 
 const content: Record<Locale, { title: string }> = {
   es: { title: "Grano de arena" },
