+++ src/app/page.tsx
@@ -9,10 +9,16 @@ export default function RootRedirect() {
   }, []);
 
   return (
-    <main style={{ padding: "2rem", textAlign: "center" }}>
-      <p>
-        <a href={`/${defaultLocale}`}>OAN International</a>
-      </p>
-    </main>
+    <>
+      {/* Crawlers / no-JS clients: 0s meta refresh (React 19 hoists this to
+          <head>). JS clients are redirected instantly by the effect above. */}
+      <meta httpEquiv="refresh" content={`0; url=/${defaultLocale}`} />
+      <link rel="canonical" href={`/${defaultLocale}`} />
+      <main style={{ padding: "2rem", textAlign: "center" }}>
+        <p>
+          <a href={`/${defaultLocale}`}>OAN International</a>
+        </p>
+      </main>
+    </>
   );
 }
