+++ src/app/sitemap.ts
@@ -1,0 +1,36 @@
+import type { MetadataRoute } from "next";
+import { locales } from "@/i18n/config";
+import { SITE_URL, STATIC_ROUTES } from "@/i18n/routes";
+import { BLOG_ARTICLES } from "@/constants/news/blog-articles";
+
+// Static export requires this so the sitemap is emitted as a static file.
+export const dynamic = "force-static";
+
+function url(locale: string, path: string): string {
+  return path ? `${SITE_URL}/${locale}/${path}` : `${SITE_URL}/${locale}`;
+}
+
+// Build the hreflang alternates map for a given locale-agnostic path.
+function languagesFor(path: string): Record<string, string> {
+  return Object.fromEntries(locales.map((l) => [l, url(l, path)]));
+}
+
+export default function sitemap(): MetadataRoute.Sitemap {
+  const paths: string[] = [
+    ...STATIC_ROUTES,
+    // Blog article ids are identical across locales.
+    ...BLOG_ARTICLES.es.map((a) => `news/${a.id}`),
+  ];
+
+  const entries: MetadataRoute.Sitemap = [];
+  for (const path of paths) {
+    const languages = languagesFor(path);
+    for (const locale of locales) {
+      entries.push({
+        url: url(locale, path),
+        alternates: { languages },
+      });
+    }
+  }
+  return entries;
+}
