+++ src/constants/membership.ts
@@ -1,0 +1,6 @@
+// Single source of truth for the members counter shown across the site
+// (home hero, become-partner, MembershipGoals). Update here only.
+export const MEMBERSHIP = {
+  current: 84,
+  target: 100,
+} as const;
